Fix: Correct dependency installation and resolve test failures (#60)
This commit addresses the GitHub Actions CI failures caused by `pytest` not being found in the PATH.

The key changes include:
- Moving the development dependencies from `[project.optional-dependencies]` to `[tool.pdm.dev-dependencies]` in `pyproject.toml`. This ensures that `pdm install -d` correctly installs `pytest` and other development tools.
- Updating `tests/unit/test_config.py` to use `monkeypatch` to unset environment variables during a specific test, making the test more robust and isolated from the execution environment.
- Creating a `tests/unit/conftest.py` file and moving the shared `sample_dsd` fixture into it. This resolves a `fixture not found` error and improves the structure of the test suite.
- Updating the `pdm.lock` file to reflect the changes in `pyproject.toml`.

With these changes, all tests now pass successfully in the CI environment.
